+++ src/neptune/management/exceptions.py
@@ -201,3 +201,12 @@ class ProjectPrivacyRestrictedException(ManagementOperationFailure):
         else:
             modified_kwargs["requested"] = '"' + requested + '"'
         super().__init__(**modified_kwargs)
+
+
+class ActiveProjectsLimitReachedException(ManagementOperationFailure):
+    code = 26
+    description = (
+        "Limit of active projects reached. You can have up to {currentQuota} active projects simultaneously. "
+        "To create a new project, you need to either archive an active project or increase the quota of active "
+        "projects in the workspace."
+    )